requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiogram==2.25.1
aiogram-di==0.1.0
aiogram-utils==0.1.4
aiogram_tools==1.0.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
crontab==1.0.1
dnspython==2.4.2
env==0.1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
magic-filter==1.0.11
mongo==0.2.0
mongoengine==0.24.2
motor==3.3.1
multidict==4.7.6
outcome==1.2.0
packaging==23.1
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
PySocks==1.7.1
pyTelegramBotAPI==4.13.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
telebot==0.0.5
telegram==0.0.1
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
